requirements.txt
torch
trimesh
PyYAML
numpy
scikit-image
h5py
opencv-python
ConfigArgParse
tensorboard
tqdm